+++ src/apis/authorization.js
@@ -3,7 +3,7 @@ const getToken = () => localStorage.getItem('token')
 
 export default {
   // 帶入需要的參數
-  loginIn({ account, password }) {
+  login({ account, password }) {
     // 這裡 return 的會是一個 Promise
     return apiHelper.post('/users/signin', {
       account,
